// 02_Template/src/app/Interfaces/icliente.ts
//BBDD_field names
export interface ICliente {
  idClients: number;
  client_name: string;
  client_surename: string;
  client_phonenumber: string;
  client_email: string;
}

// export interface ICliente {
//   idClientes: number;
//   Nombres: string;
//   Direccion: string;
//   Telefono: string;
//   Cedula: string;
//   Correo: string;
// }